Servant.Client

Description

This module provides client which can automatically generate querying functions for each endpoint just from the type representing your API.

Documentation

client :: HasClient layout => Proxy layout -> Client layout

client allows you to produce operations to query an API from a client.

type MyApi = "books" :> Get [Book] -- GET /books
        :<|> "books" :> ReqBody Book :> Post Book -- POST /books

myApi :: Proxy MyApi
myApi = Proxy

getAllBooks :: BaseUrl -> EitherT String IO [Book]
postNewBook :: Book -> BaseUrl -> EitherT String IO Book
(getAllBooks :<|> postNewBook) = client myApi

class HasClient layout where

This class lets us define how each API combinator influences the creation of an HTTP request. It's mostly an internal class, you can just use client.

Associated Types

type Client layout :: *

Methods

clientWithRoute :: Proxy layout -> Req -> Client layout

module Servant.Common.BaseUrl