Servant.Common.Req

Documentation

data ServantError

Constructors

FailureResponse

Fields

failingRequest :: UrlReq

responseStatus :: Status

responseContentType :: MediaType

responseBody :: ByteString

DecodeFailure

Fields

decodeError :: String

responseContentType :: MediaType

responseBody :: ByteString

UnsupportedContentType

Fields

responseContentType :: MediaType

responseBody :: ByteString

InvalidContentTypeHeader

Fields

responseContentTypeHeader :: ByteString

responseBody :: ByteString

ConnectionError

Fields

connectionError :: SomeException

data UrlReq

Constructors

UrlReq BaseUrl Req

data Req

Constructors

Req

Fields

reqPath :: Builder

qs :: QueryText

reqBody :: Maybe (RequestBody, MediaType)

reqAccept :: [MediaType]

headers :: [(String, Text)]

defReq :: Req

appendToPath :: String -> Req -> Req

appendToQueryString

Arguments

:: Text
param name

-> Maybe Text
param value

-> Req

-> Req

addHeader :: ToHttpApiData a => String -> a -> Req -> Req

setRQBody :: ByteString -> MediaType -> Req -> Req

Deprecated: Use setReqBodyLBS instead

Set body and media type of the request being constructed.

The body is set to the given bytestring using the RequestBodyLBS constructor.

setReqBodyLBS :: ByteString -> MediaType -> Req -> Req

Set body and media type of the request being constructed.

The body is set to the given bytestring using the RequestBodyLBS constructor.

Since: 0.9.2.0

setReqBody :: RequestBody -> MediaType -> Req -> Req

Set body and media type of the request being constructed.

Since: 0.9.2.0

reqToRequest :: (Functor m, MonadThrow m) => Req -> BaseUrl -> m Request

performing requests

displayHttpRequest :: Method -> String

data ClientEnv

Constructors

ClientEnv

Fields

manager :: Manager

baseUrl :: BaseUrl

newtype ClientM a

ClientM is the monad in which client functions run. Contains the Manager and BaseUrl used for requests in the reader environment.

Constructors

ClientM

Fields

runClientM' :: ReaderT ClientEnv (ExceptT ServantError IO) a

runClientM :: ClientM a -> ClientEnv -> IO (Either ServantError a)

performRequest :: Method -> Req -> ClientM (Int, ByteString, MediaType, [Header], Response ByteString)

performRequestCT :: MimeUnrender ct result => Proxy ct -> Method -> Req -> ClientM ([Header], result)

performRequestNoBody :: Method -> Req -> ClientM [Header]

catchConnectionError :: IO a -> IO (Either ServantError a)