Servant.Client

Description

This module provides client which can automatically generate querying functions for each endpoint just from the type representing your API.

Documentation

type family AuthClientData a :: *

For a resource protected by authentication (e.g. AuthProtect), we need to provide the client with some data used to add authentication data to a request

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

newtype AuthenticateReq a

For better type inference and to avoid usage of a data family, we newtype wrap the combination of some AuthClientData and a function to add authentication data to a request

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

Constructors

AuthenticateReq

Fields

unAuthReq :: (AuthClientData a, AuthClientData a -> Req -> Req)

client :: HasClient api => Proxy api -> Client api

client allows you to produce operations to query an API from a client.

type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
        :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book -- POST /books

myApi :: Proxy MyApi
myApi = Proxy

getAllBooks :: ClientM [Book]
postNewBook :: Book -> ClientM Book
(getAllBooks :<|> postNewBook) = client myApi

class HasClient api where

This class lets us define how each API combinator influences the creation of an HTTP request. It's mostly an internal class, you can just use client.

Minimal complete definition

clientWithRoute

Associated Types

type Client api :: *

Methods

clientWithRoute :: Proxy api -> Req -> Client api

data ClientM a

ClientM is the monad in which client functions run. Contains the Manager and BaseUrl used for requests in the reader environment.

runClientM :: ClientM a -> ClientEnv -> IO (Either ServantError a)

data ClientEnv

Constructors

ClientEnv Manager BaseUrl

mkAuthenticateReq :: AuthClientData a -> (AuthClientData a -> Req -> Req) -> AuthenticateReq a

Handy helper to avoid wrapping datatypes in tuples everywhere.

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

data ServantError

Constructors

FailureResponse

Fields

failingRequest :: UrlReq

responseStatus :: Status

responseContentType :: MediaType

responseBody :: ByteString

DecodeFailure

Fields

decodeError :: String

responseContentType :: MediaType

responseBody :: ByteString

UnsupportedContentType

Fields

responseContentType :: MediaType

responseBody :: ByteString

InvalidContentTypeHeader

Fields

responseContentTypeHeader :: ByteString

responseBody :: ByteString

ConnectionError

Fields

connectionError :: SomeException

data EmptyClient

Singleton type representing a client for an empty API.

Constructors

EmptyClient

module Servant.Common.BaseUrl