Servant.Client

Description

This module provides client which can automatically generate querying functions for each endpoint just from the type representing your API.

Documentation

client :: HasClient ClientM api => Proxy api -> Client ClientM api

Generates a set of client functions for an API.

Example:

type API = Capture "no" Int :> Get '[JSON] Int
       :<|> Get '[JSON] [Bool]

api :: Proxy API
api = Proxy

getInt :: Int -> ClientM Int
getBools :: ClientM [Bool]
getInt :<|> getBools = client api

data ClientM a

ClientM is the monad in which client functions run. Contains the Manager and BaseUrl used for requests in the reader environment.

runClientM :: ClientM a -> ClientEnv -> IO (Either ClientError a)

data ClientEnv

The environment in which a request is run. The baseUrl and makeClientRequest function are used to create a http-client request. Cookies are then added to that request if a CookieJar is set on the environment. Finally the request is executed with the manager. The makeClientRequest function can be used to modify the request to execute and set values which are not specified on a servant RequestF like responseTimeout or redirectCount

Constructors

ClientEnv

Fields

manager :: Manager

baseUrl :: BaseUrl

cookieJar :: Maybe (TVar CookieJar)

makeClientRequest :: BaseUrl -> Request -> IO Request

this function can be used to customize the creation of http-client requests from servant requests. Default value: defaultMakeClientRequest Note that: 1. makeClientRequest exists to allow overriding operational semantics e.g. responseTimeout per request, If you need global modifications, you should use managerModifyRequest 2. the cookieJar, if defined, is being applied after makeClientRequest is called.

mkClientEnv :: Manager -> BaseUrl -> ClientEnv

ClientEnv smart constructor.

defaultMakeClientRequest :: BaseUrl -> Request -> IO Request

Create a http-client Request from a servant RequestF The host, path and port fields are extracted from the BaseUrl otherwise the body, headers and query string are derived from the servant RequestF

hoistClient :: HasClient ClientM api => Proxy api -> (forall a. m a -> n a) -> Client m api -> Client n api

Change the monad the client functions live in, by supplying a conversion function (a natural transformation to be precise).

For example, assuming you have some manager :: Manager and baseurl :: BaseUrl around:

type API = Get '[JSON] Int :<|> Capture "n" Int :> Post '[JSON] Int
api :: Proxy API
api = Proxy
getInt :: IO Int
postInt :: Int -> IO Int
getInt :<|> postInt = hoistClient api (flip runClientM cenv) (client api)
  where cenv = mkClientEnv manager baseurl

module Servant.Client.Core.Reexport