Servant.Client.Internal.HttpClient

Documentation

data ClientEnv

The environment in which a request is run.

Constructors

ClientEnv

Fields

manager :: Manager

baseUrl :: BaseUrl

cookieJar :: Maybe (TVar CookieJar)

mkClientEnv :: Manager -> BaseUrl -> ClientEnv

ClientEnv smart constructor.

client :: HasClient ClientM api => Proxy api -> Client ClientM api

Generates a set of client functions for an API.

Example:

type API = Capture "no" Int :> Get '[JSON] Int
       :<|> Get '[JSON] [Bool]

api :: Proxy API
api = Proxy

getInt :: Int -> ClientM Int
getBools :: ClientM [Bool]
getInt :<|> getBools = client api

hoistClient :: HasClient ClientM api => Proxy api -> (forall a. m a -> n a) -> Client m api -> Client n api

Change the monad the client functions live in, by supplying a conversion function (a natural transformation to be precise).

For example, assuming you have some manager :: Manager and baseurl :: BaseUrl around:

type API = Get '[JSON] Int :<|> Capture "n" Int :> Post '[JSON] Int
api :: Proxy API
api = Proxy
getInt :: IO Int
postInt :: Int -> IO Int
getInt :<|> postInt = hoistClient api (flip runClientM cenv) (client api)
  where cenv = mkClientEnv manager baseurl

newtype ClientM a

ClientM is the monad in which client functions run. Contains the Manager and BaseUrl used for requests in the reader environment.

Constructors

ClientM

Fields

unClientM :: ReaderT ClientEnv (ExceptT ServantError IO) a

runClientM :: ClientM a -> ClientEnv -> IO (Either ServantError a)

performRequest :: Request -> ClientM Response

clientResponseToResponse :: Response a -> GenResponse a

requestToClientRequest :: BaseUrl -> Request -> Request

catchConnectionError :: IO a -> IO (Either ServantError a)